Fix 32-bit overflow with --no-net
--no-net causes tarballTtl to be set to the largest 32-bit integer,
which causes comparison like 'time + tarballTtl < other_time' to
fail on 32-bit systems. So cast them to 64-bit first.

https://hydra.nixos.org/build/95076624

+++ src/libexpr/primops/fetchGit.cc
@@ -135,7 +135,7 @@ GitInfo exportGit(ref<Store> store, std::string uri,
                git fetch to update the local ref to the remote ref. */
             struct stat st;
             doFetch = stat(localRefFile.c_str(), &st) != 0 ||
-                st.st_mtime + settings.tarballTtl <= now;
+                (uint64_t) st.st_mtime + settings.tarballTtl <= (uint64_t) now;
         }
 
         if (doFetch) {
